requirements.txt
pre-commit
numpy
pandas
matplotlib
seaborn
torch
lightning
torchdata
wandb
omegaconf
hydra-core
beartype
submitit
hydra-submitit-launcher
pytest
einops
scipy
transformers
datasets